nexus/requirements.txt
streamlit
plotly
pandas
redis
numpy
streamlit-autorefresh
opencv-python-headless
pytesseract
pyserial
Pillow
